add access.log

+++ index.py
@@ -1,9 +1,14 @@
 #! /usr/bin/env python
 
-from bottle import route, run, response
+from bottle import Bottle, run, response
 import paramiko
 import subprocess
 import configparser
+from requestlogger import WSGILogger, ApacheFormatter
+from logging.handlers import TimedRotatingFileHandler
+import os
+
+app = Bottle()
 
 config = configparser.ConfigParser()
 config.read('config.ini')
@@ -29,28 +34,37 @@ API_KEY = ConfigSectionMap("MAIN")["api_key"]
 def send_wol(ip, mac):
     p = subprocess.Popen("wakeonlan -i %s -p 22 %s" % (ip, mac), shell=True, stdout=subprocess.PIPE,
                          stderr=subprocess.STDOUT)
+
     return p.stdout.readlines()
 
 
 def send_shutdown(ip, user, pw):
-    ssh = paramiko.SSHClient()
-    ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
-    try:
-        ssh.connect(ip, username=user, password=pw)
-    except paramiko.ssh_exception.NoValidConnectionsError:
-        return "already shutdown"
 
-    ssh.exec_command("shutdown -p now")
-    return "shutting down NAS"
+    nas_up = True if os.system("ping -c 1 " + ip) is 0 else False
+
+    if nas_up:
+
+        ssh = paramiko.SSHClient()
+        ssh.set_missing_host_key_policy(paramiko.AutoAddPolicy())
+        try:
+            ssh.connect(ip, username=user, password=pw)
+        except paramiko.ssh_exception.NoValidConnectionsError:
+            return "No SSH Connection"
+
+        ssh.exec_command("shutdown -p now")
+
+        return "shutting down NAS"
+    else:
+        return "already down"
 
 
-@route('/')
+@app.route('/')
 def default_route():
     response.status = 403
     return 'Forbidden'
 
 
-@route('/start-nas/<key>')
+@app.route('/start-nas/<key>')
 def start_nas(key):
     if key != API_KEY:
         response.status = 401
@@ -60,7 +74,7 @@ def start_nas(key):
         return send_wol(ConfigSectionMap("NAS")["ip"], ConfigSectionMap("NAS")["mac"])
 
 
-@route('/stop-nas/<key>')
+@app.route('/stop-nas/<key>')
 def stop_nas(key):
     if key != API_KEY:
         response.status = 401
@@ -70,8 +84,10 @@ def stop_nas(key):
         return send_shutdown(ConfigSectionMap("NAS")["ip"], ConfigSectionMap("NAS")["ssh_user"],
                              ConfigSectionMap("NAS")["ssh_pass"])
 
+handlers = [TimedRotatingFileHandler('access.log', 'd', 7), ]
+app = WSGILogger(app, handlers, ApacheFormatter(), ip_header='HTTP_X_FORWARDED_FOR')
 
-run(host=ConfigSectionMap("MAIN")["host"], port=ConfigSectionMap("MAIN")["port"],
+run(app, host=ConfigSectionMap("MAIN")["host"], port=ConfigSectionMap("MAIN")["port"],
     debug=ConfigSectionMap("MAIN")["debug"])
